ChatGPT help
ChatGPT recommended removing the if-else statement at the end of the function and adding a for-else statement.  However, after this change, I no longer needed the 'elif num == 2' test.

The actual recommendation was this:
def is_prime_number(num):
    if num <= 1:
        return False
    elif num == 2:
        return True
    else:
        num = int(num)
        for n in range(2, num):
            if num % n == 0:
                return False
        return True

As the application worked before and after in PE q17 and q18, I am only changing q17 to reflect this recommendation.